refactor: code refactoring(remove duplicate method)
