+++ apps/calendar/src/components/panelgrid/panelEvents.tsx
@@ -3,12 +3,12 @@ import { FunctionComponent, h } from 'preact';
 import { cls } from '@src/util/cssHelper';
 import Schedule from '@src/model/schedule';
 import { DayEvent } from '@src/components/events/dayEvent';
-import { EVENT_HEIGHT, getViewModels } from '@src/event/panelEvent';
+import { getViewModels } from '@src/event/panelEvent';
 import { setViewModelsInfo } from '@src/event/gridEvent';
-import ScheduleViewModel from '@src/model/scheduleViewModel';
 
 import type { Cells } from '@t/panel';
 import type { PanelState } from '@src/components/layout';
+import { EVENT_HEIGHT, isWithinHeight } from '@src/util/gridHelper';
 
 interface Props {
   name: string;
@@ -16,10 +16,6 @@ interface Props {
   events: Schedule[];
   options?: PanelState;
 }
-
-const isWithinHeight = (panelHeight: number, eventHeight: number) => {
-  return ({ top }: ScheduleViewModel) => panelHeight >= (top + 1) * eventHeight;
-};
 
 export const PanelEvents: FunctionComponent<Props> = ({ name, cells, events, options = {} }) => {
   const { narrowWeekend = false, panelHeight = EVENT_HEIGHT } = options;
