+++ apps/calendar/test/app/time/datetime.spec.ts
@@ -515,4 +515,50 @@ describe('datetime', () => {
 
     expect(dt.toEndOfYear(date)).toEqual(endOfYear);
   });
+
+  describe('getMonthCalendar', () => {
+    const date = createDate(2021, 5, 14);
+
+    it("should make a month with starting from Wednesday, If 'startDayOfWeek' is 3", () => {
+      expect(dt.getMonthCalendar(date, { startDayOfWeek: 3, visibleWeeksCount: 2 })).toEqual([
+        [
+          createDate(2021, 5, 12),
+          createDate(2021, 5, 13),
+          createDate(2021, 5, 14),
+          createDate(2021, 5, 15),
+          createDate(2021, 5, 16),
+          createDate(2021, 5, 17),
+          createDate(2021, 5, 18),
+        ],
+        [
+          createDate(2021, 5, 19),
+          createDate(2021, 5, 20),
+          createDate(2021, 5, 21),
+          createDate(2021, 5, 22),
+          createDate(2021, 5, 23),
+          createDate(2021, 5, 24),
+          createDate(2021, 5, 25),
+        ],
+      ]);
+    });
+
+    it("should return a month only weekdays, If 'workweek' is true", () => {
+      expect(dt.getMonthCalendar(date, { workweek: true, visibleWeeksCount: 2 })).toEqual([
+        [
+          createDate(2021, 5, 10),
+          createDate(2021, 5, 11),
+          createDate(2021, 5, 12),
+          createDate(2021, 5, 13),
+          createDate(2021, 5, 14),
+        ],
+        [
+          createDate(2021, 5, 17),
+          createDate(2021, 5, 18),
+          createDate(2021, 5, 19),
+          createDate(2021, 5, 20),
+          createDate(2021, 5, 21),
+        ],
+      ]);
+    });
+  });
 });
