+++ apps/calendar/types/components/events/gridEvent.d.ts
@@ -1,0 +1,7 @@
+type EventItemStyleParam = {
+  flat: boolean;
+  bgColor: string;
+  borderColor: string;
+  exceedLeft: boolean;
+  exceedRight: boolean;
+};
